Update MyDownloads coremenuoption. Add temporary checks to account for differnt route structure on MyDownloads SPA. Will need to be updated to be more flexible
